Always include FN field in generated vCard, even if empty
According to the specification
( https://tools.ietf.org/html/rfc6350#section-6.2.1 ),
FN is a required field.

Generating a vCard without an FN field could lead to the other
side rejecting it as invalid.

This is the case with Nextcloud Contacts' import service
and causes problems like nextcloud/contacts#413
( https://github.com/nextcloud/contacts/issues/413 ).
Nextcloud Contacts makes use of this library to preprocess
`.vcf` files. Parsing entries with empty FN values is handled well
by this library. But as later on, new vCard entries are
generated (`vCard.generate()`) out of that data,
which completely lack an FN field.
Those FN-field-lacking vCards then reach the PHP server, which
chokes on them, as they are indeed invalid according to the specs.

This patch ensures that FN is always included in generated vCards,
even if the FN input is an empty string.

An FN value of `undefined` still yields vCards without an FN field.
That code remains unchanged. How it should be handled is a separate
issue.